Looks like that was the last real bug.
The other thing is to initialize by randomly sampling cluster
parameters, not the assignment.  Why?  Because initializing by
randomly sampling the assignment makes the resulting clusters almost
identical, and then it takes a huge number of Gibbs sweeps before they
drift apart.  But if we start with random cluster parameters, they are
already different, and then the assignment gives them distinct parts
of the data, so they can converge much faster.